requirements.txt
insightface==0.7.3
onnxruntime-gpu==1.24.1
opencv-python==4.13.0.92
numpy==2.3.3
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.19.0.56
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nvjitlink-cu12==12.9.86
scikit-learn==1.3.0
# ── FastAPI Backend ──
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
# ── FAISS (choose one) ──
# For GPU (requires CUDA):
faiss-gpu>=1.7.4
# For CPU only (if you don't have a GPU):
# faiss-cpu>=1.7.4
